+++ src/Symfony/Component/Cache/DependencyInjection/CachePoolPass.php
@@ -50,7 +50,7 @@ class CachePoolPass implements CompilerPassInterface
         if ($container->hasParameter('cache.prefix.seed')) {
             $seed = '.'.$container->getParameterBag()->resolveValue($container->getParameter('cache.prefix.seed'));
         } else {
-            $seed = '_'.$container->getParameter('kernel.root_dir');
+            $seed = '_'.$container->getParameter('kernel.project_dir');
         }
         $seed .= '.'.$container->getParameter('kernel.container_class').'.'.$container->getParameter('kernel.environment');